requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.3
cssbeautifier==1.15.4
defusedxml==0.7.1
Django==5.2
django-rest-swagger==2.2.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djlint==1.36.4
djoser==2.3.1
drf-yasg==1.21.10
EditorConfig==0.17.0
flake8==7.2.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
itypes==1.2.0
Jinja2==3.1.6
jsbeautifier==1.15.4
json5==0.12.0
kombu==5.5.3
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
simplejson==3.20.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
gunicorn==23.0.0